Feat: SEO 친화적인 URL 패턴 구현
- 공연명-짧은ID 형태의 SEO 최적화 URL 생성
- URL 슬러그 유틸리티 함수 추가 (createSeoConcertUrl, extractConcertShortId)
- 메인 페이지에서 SEO 친화적인 URL로 공연 상세 페이지 이동
- 백엔드에서 짧은 ID로 공연 조회 기능 구현
- 새로운 공연 상세 페이지 (/concert/[slug]) 생성
- 기존 복잡한 UUID URL 대신 사용자 친화적 URL 제공
